update bai cu

+++ MTshop/database/migrations/2019_12_07_070748_create_cusc_mau_sanpham_table.php
@@ -16,7 +16,7 @@ class CreateCuscMauSanphamTable extends Migration
         Schema::create('cusc_mau_sanpham', function (Blueprint $table) {
             $table->engine = 'InnoDB';
             $table->unsignedBigInteger('sp_ma')->comment('Màu sắc # m_ma # m_ten # Mã sản phẩm');
-            $table->unsignedBigInteger('m_ma')->comment('Sản phẩm # sp_ma # sp_ten # Mã màu sản phẩm');
+            $table->unsignedTinyInteger('m_ma')->comment('Sản phẩm # sp_ma # sp_ten # Mã màu sản phẩm');
             $table->unsignedSmallInteger('msp_soluong')->default('0')->comment('Số lượng # Số lượng sản phẩm tương ứng với màu');
             
             $table->primary(['sp_ma', 'm_ma']);
